C: defer REGCOMP for \p{ properties
add a re_does_swash() detector (sans the utf8 case folding).
the re extflags cannot tell use precisely when a regex will need
a swash_init, which needs to be deferred to init1.
so check manually if the regex contains a unicode property syntax
\p{ or \P{.
change $sv->EXTFLAGS to compflags since 5.22 for CALLREGCOMP()

Fixes #253 for 5.24.